Custom Debug Hook in Go via "State.SetHook"
With this commit it's now possible to define a custom debug Hook in Go
via the new Function "State.SetHook". Currently there is only one Hook
per State which means that a call to "State.SetHook" will always
override the previously set hook.

This hook is always of type "LUA_MASKCOUNT"